+++ R/clusterSetup.R
@@ -285,13 +285,13 @@ clusterSetup <- function(messagePrefix = "DEoptim_",
     stMoveObjects <- try({
       system.time({
         objsToCopy <- mget(unlist(objsNeeded), envir = envir)
-        FileBackendsToCopy <- Filenames(objsToCopy)
+        FileBackendsToCopy <- reproducible::Filenames(objsToCopy)
         hasFilename <- nzchar(FileBackendsToCopy)
         if (any(hasFilename)) {
           objsToMem <- names(FileBackendsToCopy)[hasFilename]
           objsToCopy[objsToMem] <-
             lapply(objsToCopy[objsToMem],
-                   function(x) toMemory(x))
+                   function(x) terra::toMemory(x))
         }
         objsToCopy <- reproducible::.wrap(objsToCopy)
         filenameForTransfer <- normalizePath(tempfile(fileext = ".qs2"), mustWork = FALSE, winslash = "/")
@@ -340,8 +340,8 @@ clusterSetup <- function(messagePrefix = "DEoptim_",
     
     if (is(stMoveObjects, "try-error")) {
       message("The attempt to move objects to cluster using rsync and qs2 failed; trying clusterExport")
-      stMoveObjects <- system.time(parallel::clusterExport(clThird, objsNeeded, envir = environment()))
-      list2env(mget(unlist(objsNeeded), envir = environment()), envir = .GlobalEnv)
+      stMoveObjects <- system.time(parallel::clusterExport(clThird, objsNeeded, envir = envir))
+      list2env(mget(unlist(objsNeeded), envir = envir), envir = .GlobalEnv)
     }
     message("it took ", round(stMoveObjects[3], 2), "s to move objects to nodes")
     control$cluster <- clThird
@@ -493,8 +493,10 @@ summaryOutputFolder <- function(path, pattern = "^.+hists/(.+)\\_iter.+\\_[[:dig
 
 
 
-changeNodenameToLocalhost <- function(cores) {
-  sshLines <- readLines("~/.ssh/config")
+changeNodenameToLocalhost <- function(cores, sshConfig = "~/.ssh/config") {
+  ## no ssh config (a laptop, a CI runner): nothing to rename
+  if (!file.exists(sshConfig)) return(cores)
+  sshLines <- readLines(sshConfig)
   hasSelf <- grep(Sys.info()["nodename"], sshLines, value = T)
   onlyHost <- grep("^Host ", hasSelf, value = TRUE)
   whLocalhost <- gsub("^Host (\\w+).*", "\\1", onlyHost)